Move handler to pkg/component package
