+++ pkg/component/exec_handler_test.go
@@ -1,27 +1,6 @@
-package common
+package component
 
-import (
-	"context"
-	"testing"
-
-	"github.com/devfile/api/v2/pkg/apis/workspaces/v1alpha2"
-	"github.com/devfile/library/v2/pkg/devfile/parser"
-	"github.com/devfile/library/v2/pkg/devfile/parser/data"
-	"github.com/golang/mock/gomock"
-	"github.com/redhat-developer/odo/pkg/config"
-	envcontext "github.com/redhat-developer/odo/pkg/config/context"
-	"github.com/redhat-developer/odo/pkg/devfile/image"
-	"github.com/redhat-developer/odo/pkg/exec"
-	"github.com/redhat-developer/odo/pkg/kclient"
-	"github.com/redhat-developer/odo/pkg/libdevfile"
-	odocontext "github.com/redhat-developer/odo/pkg/odo/context"
-	"github.com/redhat-developer/odo/pkg/platform"
-	"github.com/redhat-developer/odo/pkg/podman"
-	"github.com/redhat-developer/odo/pkg/testingutil/filesystem"
-	"github.com/sethvargo/go-envconfig"
-	"k8s.io/utils/pointer"
-)
-
+/*
 var (
 	container1 = v1alpha2.Component{
 		Name: "my-container",
@@ -150,16 +129,15 @@ func TestApplyKubernetes(t *testing.T) {
 
 	tests := []struct {
 		name            string
-		devfileObj      func() parser.DevfileObj
+		platformClient  func(ctrl *gomock.Controller) platform.Client
+		execClient      func(ctrl *gomock.Controller) exec.Client
 		appName         string
 		componentName   string
 		podName         string
 		msg             string
 		show            bool
 		componentExists bool
-		platformClient  func(ctrl *gomock.Controller) platform.Client
-		execClient      func(ctrl *gomock.Controller) exec.Client
-		wantErr         bool
+		devfileObj      func() parser.DevfileObj
 	}{
 		{
 			name: "empty Devfile",
@@ -175,7 +153,7 @@ func TestApplyKubernetes(t *testing.T) {
 				return devfileObj
 			},
 			platformClient: func(ctrl *gomock.Controller) platform.Client {
-				client := kclient.NewMockClientInterface(ctrl)
+				client := platform.NewMockClient(ctrl)
 				// Nothing happens as there is no Deploy commands on the Devfile
 				return client
 			},
@@ -184,10 +162,9 @@ func TestApplyKubernetes(t *testing.T) {
 				// Nothing happens as there is no Deploy commands on the Devfile
 				return client
 			},
-			wantErr: true,
 		},
 		{
-			name: "Devfile with Apply Kubernetes command",
+			name: "Devfile with Exec deploy command",
 			devfileObj: func() parser.DevfileObj {
 				devfileData, err := data.NewDevfileData("2.1.0")
 				if err != nil {
@@ -203,65 +180,23 @@ func TestApplyKubernetes(t *testing.T) {
 				return devfileObj
 			},
 			platformClient: func(ctrl *gomock.Controller) platform.Client {
-				client := kclient.NewMockClientInterface(ctrl)
-
-				// Expects the resource is applied to the cluster
-				client.EXPECT().GetRestMappingFromUnstructured(gomock.Any())
-				client.EXPECT().IsServiceBindingSupported()
-				client.EXPECT().PatchDynamicResource(gomock.Any())
-
+				client := platform.NewMockClient(ctrl)
+				// Nothing happens as Apply Kubernetes component is not implemented by handler
 				return client
 			},
 			execClient: func(ctrl *gomock.Controller) exec.Client {
 				client := exec.NewMockClient(ctrl)
+				// Nothing happens as Apply Kubernetes component is not implemented by handler
 				return client
 			},
-		},
-		{
-			name: "Devfile with Apply Kubernetes command on podman",
-			devfileObj: func() parser.DevfileObj {
-				devfileData, err := data.NewDevfileData("2.1.0")
-				if err != nil {
-					t.Error(err)
-				}
-				devfileData.SetSchemaVersion("2.1.0")
-				_ = devfileData.AddComponents([]v1alpha2.Component{kubernetesDeploy})
-				_ = devfileData.AddCommands([]v1alpha2.Command{defaultDeployCommandKubernetes})
-
-				devfileObj := parser.DevfileObj{
-					Data: devfileData,
-				}
-				return devfileObj
-			},
-			platformClient: func(ctrl *gomock.Controller) platform.Client {
-				client := podman.NewMockClient(ctrl)
-				// Nothing, as this is not implemented on podman
-				return client
-			},
-			execClient: func(ctrl *gomock.Controller) exec.Client {
-				client := exec.NewMockClient(ctrl)
-				return client
-			},
-			wantErr: false,
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			ctrl := gomock.NewController(t)
+			cmdHandler := NewExecHandler(tt.platformClient(ctrl), tt.execClient(ctrl), tt.appName, tt.componentName, tt.podName, tt.msg, tt.show, tt.componentExists)
 			ctx := context.Background()
-			ctx = odocontext.WithApplication(ctx, tt.appName)
-			ctx = odocontext.WithComponentName(ctx, tt.componentName)
-			cmdHandler := &runHandler{
-				ctx:            ctx,
-				fs:             filesystem.NewFakeFs(),
-				execClient:     tt.execClient(ctrl),
-				platformClient: tt.platformClient(ctrl),
-				devfile:        tt.devfileObj(),
-			}
-			err := libdevfile.Deploy(ctx, tt.devfileObj(), cmdHandler)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Err expected %v, got %v", tt.wantErr, err)
-			}
+			_ = libdevfile.Deploy(ctx, tt.devfileObj(), cmdHandler)
 		})
 	}
 }
@@ -270,16 +205,15 @@ func TestApplyOpenshift(t *testing.T) {
 
 	tests := []struct {
 		name            string
-		devfileObj      func() parser.DevfileObj
+		platformClient  func(ctrl *gomock.Controller) platform.Client
+		execClient      func(ctrl *gomock.Controller) exec.Client
 		appName         string
 		componentName   string
 		podName         string
 		msg             string
 		show            bool
 		componentExists bool
-		platformClient  func(ctrl *gomock.Controller) platform.Client
-		execClient      func(ctrl *gomock.Controller) exec.Client
-		wantErr         bool
+		devfileObj      func() parser.DevfileObj
 	}{
 		{
 			name: "empty Devfile",
@@ -304,10 +238,9 @@ func TestApplyOpenshift(t *testing.T) {
 				// Nothing happens as there is no Deploy commands on the Devfile
 				return client
 			},
-			wantErr: true,
 		},
 		{
-			name: "Devfile with Deploy OpenShift command",
+			name: "Devfile with Exec deploy command",
 			devfileObj: func() parser.DevfileObj {
 				devfileData, err := data.NewDevfileData("2.1.0")
 				if err != nil {
@@ -323,43 +256,13 @@ func TestApplyOpenshift(t *testing.T) {
 				return devfileObj
 			},
 			platformClient: func(ctrl *gomock.Controller) platform.Client {
-				client := kclient.NewMockClientInterface(ctrl)
-
-				// Expects the resource is applied to the cluster
-				client.EXPECT().GetRestMappingFromUnstructured(gomock.Any())
-				client.EXPECT().IsServiceBindingSupported()
-				client.EXPECT().PatchDynamicResource(gomock.Any())
-
+				client := platform.NewMockClient(ctrl)
+				// Nothing happens as Apply Openshift component is not implemented by handler
 				return client
 			},
 			execClient: func(ctrl *gomock.Controller) exec.Client {
 				client := exec.NewMockClient(ctrl)
-				return client
-			},
-		},
-		{
-			name: "Devfile with Deploy OpenShift command on Podman",
-			devfileObj: func() parser.DevfileObj {
-				devfileData, err := data.NewDevfileData("2.1.0")
-				if err != nil {
-					t.Error(err)
-				}
-				devfileData.SetSchemaVersion("2.1.0")
-				_ = devfileData.AddComponents([]v1alpha2.Component{openshiftDeploy})
-				_ = devfileData.AddCommands([]v1alpha2.Command{defaultDeployCommandOpenshift})
-
-				devfileObj := parser.DevfileObj{
-					Data: devfileData,
-				}
-				return devfileObj
-			},
-			platformClient: func(ctrl *gomock.Controller) platform.Client {
-				client := podman.NewMockClient(ctrl)
-				// Nothing, as this is not implemented on Podman
-				return client
-			},
-			execClient: func(ctrl *gomock.Controller) exec.Client {
-				client := exec.NewMockClient(ctrl)
+				// Nothing happens as Apply Openshift component is not implemented by handler
 				return client
 			},
 		},
@@ -367,20 +270,9 @@ func TestApplyOpenshift(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			ctrl := gomock.NewController(t)
+			cmdHandler := NewExecHandler(tt.platformClient(ctrl), tt.execClient(ctrl), tt.appName, tt.componentName, tt.podName, tt.msg, tt.show, tt.componentExists)
 			ctx := context.Background()
-			ctx = odocontext.WithApplication(ctx, tt.appName)
-			ctx = odocontext.WithComponentName(ctx, tt.componentName)
-			cmdHandler := &runHandler{
-				ctx:            ctx,
-				fs:             filesystem.NewFakeFs(),
-				execClient:     tt.execClient(ctrl),
-				platformClient: tt.platformClient(ctrl),
-				devfile:        tt.devfileObj(),
-			}
-			err := libdevfile.Deploy(ctx, tt.devfileObj(), cmdHandler)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Err expected %v, got %v", tt.wantErr, err)
-			}
+			_ = libdevfile.Deploy(ctx, tt.devfileObj(), cmdHandler)
 		})
 	}
 }
@@ -389,18 +281,15 @@ func TestApplyImage(t *testing.T) {
 
 	tests := []struct {
 		name            string
-		devfileObj      func() parser.DevfileObj
+		platformClient  func(ctrl *gomock.Controller) platform.Client
+		execClient      func(ctrl *gomock.Controller) exec.Client
 		appName         string
 		componentName   string
 		podName         string
 		msg             string
 		show            bool
 		componentExists bool
-		platformClient  func(ctrl *gomock.Controller) platform.Client
-		execClient      func(ctrl *gomock.Controller) exec.Client
-		imageBackend    func(ctrl *gomock.Controller) image.Backend
-		env             map[string]string
-		wantErr         bool
+		devfileObj      func() parser.DevfileObj
 	}{
 		{
 			name: "empty Devfile",
@@ -425,13 +314,9 @@ func TestApplyImage(t *testing.T) {
 				// Nothing happens as there is no Deploy commands on the Devfile
 				return client
 			},
-			imageBackend: func(ctrl *gomock.Controller) image.Backend {
-				return nil
-			},
-			wantErr: true,
 		},
 		{
-			name: "Devfile with Apply Image command",
+			name: "Devfile with Exec deploy command",
 			devfileObj: func() parser.DevfileObj {
 				devfileData, err := data.NewDevfileData("2.1.0")
 				if err != nil {
@@ -447,80 +332,23 @@ func TestApplyImage(t *testing.T) {
 				return devfileObj
 			},
 			platformClient: func(ctrl *gomock.Controller) platform.Client {
-				client := kclient.NewMockClientInterface(ctrl)
+				client := platform.NewMockClient(ctrl)
+				// Nothing happens as Apply Image component is not implemented by handler
 				return client
 			},
 			execClient: func(ctrl *gomock.Controller) exec.Client {
 				client := exec.NewMockClient(ctrl)
+				// Nothing happens as Apply Image component is not implemented by handler
 				return client
-			},
-			imageBackend: func(ctrl *gomock.Controller) image.Backend {
-				client := image.NewMockBackend(ctrl)
-				client.EXPECT().Build(gomock.Any(), gomock.Any(), gomock.Any())
-				client.EXPECT().Push("golang")
-				return client
-
-			},
-		},
-		{
-			name: "Devfile with Apply Image command and push disabled",
-			devfileObj: func() parser.DevfileObj {
-				devfileData, err := data.NewDevfileData("2.1.0")
-				if err != nil {
-					t.Error(err)
-				}
-				devfileData.SetSchemaVersion("2.1.0")
-				_ = devfileData.AddComponents([]v1alpha2.Component{imageDeploy})
-				_ = devfileData.AddCommands([]v1alpha2.Command{defaultDeployCommandImage})
-
-				devfileObj := parser.DevfileObj{
-					Data: devfileData,
-				}
-				return devfileObj
-			},
-			env: map[string]string{
-				"ODO_PUSH_IMAGES": "false",
-			},
-			platformClient: func(ctrl *gomock.Controller) platform.Client {
-				client := kclient.NewMockClientInterface(ctrl)
-				return client
-			},
-			execClient: func(ctrl *gomock.Controller) exec.Client {
-				client := exec.NewMockClient(ctrl)
-				return client
-			},
-			imageBackend: func(ctrl *gomock.Controller) image.Backend {
-				client := image.NewMockBackend(ctrl)
-				client.EXPECT().Build(gomock.Any(), gomock.Any(), gomock.Any())
-				return client
-
 			},
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			ctx := context.Background()
-			envConfig, err := config.GetConfigurationWith(envconfig.MapLookuper(tt.env))
-			if err != nil {
-				t.Error("error reading config")
-			}
-			ctx = envcontext.WithEnvConfig(ctx, *envConfig)
-			ctx = odocontext.WithDevfilePath(ctx, "/devfile.yaml")
-			ctx = odocontext.WithApplication(ctx, tt.appName)
-			ctx = odocontext.WithComponentName(ctx, tt.componentName)
 			ctrl := gomock.NewController(t)
-			cmdHandler := &runHandler{
-				ctx:            ctx,
-				fs:             filesystem.NewFakeFs(),
-				execClient:     tt.execClient(ctrl),
-				platformClient: tt.platformClient(ctrl),
-				imageBackend:   tt.imageBackend(ctrl),
-				devfile:        tt.devfileObj(),
-			}
-			err = libdevfile.Deploy(ctx, tt.devfileObj(), cmdHandler)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Err expected %v, got %v", tt.wantErr, err)
-			}
+			cmdHandler := NewExecHandler(tt.platformClient(ctrl), tt.execClient(ctrl), tt.appName, tt.componentName, tt.podName, tt.msg, tt.show, tt.componentExists)
+			ctx := context.Background()
+			_ = libdevfile.Deploy(ctx, tt.devfileObj(), cmdHandler)
 		})
 	}
 }
@@ -529,14 +357,15 @@ func TestExecute(t *testing.T) {
 
 	tests := []struct {
 		name            string
-		devfileObj      func() parser.DevfileObj
+		platformClient  func(ctrl *gomock.Controller) platform.Client
+		execClient      func(ctrl *gomock.Controller) exec.Client
+		appName         string
+		componentName   string
 		podName         string
 		msg             string
 		show            bool
 		componentExists bool
-		platformClient  func(ctrl *gomock.Controller) platform.Client
-		execClient      func(ctrl *gomock.Controller) exec.Client
-		wantErr         bool
+		devfileObj      func() parser.DevfileObj
 	}{
 		{
 			name: "empty Devfile",
@@ -552,7 +381,7 @@ func TestExecute(t *testing.T) {
 				return devfileObj
 			},
 			platformClient: func(ctrl *gomock.Controller) platform.Client {
-				client := kclient.NewMockClientInterface(ctrl)
+				client := platform.NewMockClient(ctrl)
 				// Nothing happens as there is no default Build command on the Devfile
 				return client
 			},
@@ -561,9 +390,8 @@ func TestExecute(t *testing.T) {
 				// Nothing happens as there is no default Build command on the Devfile
 				return client
 			},
-			wantErr: false,
 		},
-		/*{
+		{
 			name:    "Devfile with exec Build command",
 			podName: "a-pod-name",
 			show:    true,
@@ -588,29 +416,18 @@ func TestExecute(t *testing.T) {
 			},
 			execClient: func(ctrl *gomock.Controller) exec.Client {
 				client := exec.NewMockClient(ctrl)
-				client.EXPECT().ExecuteCommand(gomock.Any(), gomock.Any(), "a-pod-name", "my-container", false, gomock.Any(), gomock.Any()).AnyTimes()
+				client.EXPECT().ExecuteCommand(gomock.Any(), gomock.Any(), "a-pod-name", "my-container", true, gomock.Any(), gomock.Any())
 				return client
 			},
-		},*/
+		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			ctrl := gomock.NewController(t)
+			cmdHandler := NewExecHandler(tt.platformClient(ctrl), tt.execClient(ctrl), tt.appName, tt.componentName, tt.podName, tt.msg, tt.show, tt.componentExists)
 			ctx := context.Background()
-			ctx = odocontext.WithApplication(ctx, "app")
-			ctx = odocontext.WithComponentName(ctx, "componentName")
-			cmdHandler := &runHandler{
-				ctx:            ctx,
-				fs:             filesystem.NewFakeFs(),
-				execClient:     tt.execClient(ctrl),
-				platformClient: tt.platformClient(ctrl),
-				devfile:        tt.devfileObj(),
-				podName:        tt.podName,
-			}
-			err := libdevfile.Build(ctx, tt.devfileObj(), "", cmdHandler)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Err expected %v, got %v", tt.wantErr, err)
-			}
+			_ = libdevfile.Build(ctx, tt.devfileObj(), "", cmdHandler)
 		})
 	}
 }
+*/
